+++ packages/supermassive/src/types.ts
@@ -12,7 +12,7 @@ import { ObjMap } from "./jsutils/ObjMap";
 import { Path } from "./jsutils/Path";
 import { ExecutionHooks } from "./hooks/types";
 import { FieldGroup } from "./collectFields";
-import { EncodedSchemaFragment } from "./types/definition";
+import { SchemaFragmentDefinitions } from "./schema/definition";
 
 export type ScalarTypeResolver = GraphQLScalarType;
 export type EnumTypeResolver = Record<string, unknown>;
@@ -258,7 +258,7 @@ export interface CommonExecutionArgs {
 }
 export type ExecutionWithoutSchemaArgs = CommonExecutionArgs & {
   document: DocumentNode;
-  schemaFragment: EncodedSchemaFragment; // FIXME: should be an array of fragments to be merged?
+  schemaFragment: SchemaFragmentDefinitions; // FIXME: should be an array of fragments to be merged?
 };
 
 export type ExecutionWithSchemaArgs = CommonExecutionArgs & {
